+++ packages/cli/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Major Changes
 
-- Initial release of linear-cli with core functionality and CLI interface.
+- Initial release of lnr with core functionality and CLI interface.
 
 ### Patch Changes